+++ docs/design/custom_op.md
@@ -53,6 +53,8 @@ For example:
 ```python
 --8<-- "vllm/model_executor/layers/attention/mm_encoder_attention.py:mm_encoder_attn"
 
+--8<-- "vllm/model_executor/layers/attention/rel_pos_attention.py:rel_pos_attention"
+
 --8<-- "vllm/model_executor/layers/mla.py:multi_head_latent_attention"
 ```
